// src/Redux/Enum.ts
export enum AUTH {
  SET_USER_DATA = "SET_USER_DATA",
  SET_LOGIN = "SET_LOGIN",
  GET_CAPTCHA_URL_SUCCESS = "GET_CAPTCHA_URL_SUCCESS",
}

export enum PROFILE {
  ADD_POST = "ADD-POST",
  SET_USER_PROFILE = "SET_USER_PROFILE",
  SET_STATUS = "SET_STATUS",
  DELETE_POST = "DELETE_POST",
  SAVE_PHOTO_SUCCESS = "SAVE_PHOTO_SUCCESS",
  SAVE_PROFILE_SUCCESS = "SAVE_PROFILE_SUCCESS",
  SAVE_EDIT_PROFILE_SUCCESS = "SAVE_EDIT_PROFILE_SUCCESS",
}

export enum DIALOGS {
  ADD_MESSAGE = "ADD-MESSAGE",
}

export enum USERS {
  FOLLOW = "FOLLOW",
  UNFOLLOW = "UNFOLLOW",
  SET_USERS = "SET_USERS",
  SET_CURRENT_PAGE = "SET_CURRENT_PAGE",
  SET_TOTAL_USERS_COUNT = "SET_TOTAL_USERS_COUNT",
  TOGGLE_IS_FETCHING = "TOGGLE_IS_FETCHING",
  FOLLOWING_IN_PROGRESS = "FOLLOWING_IN_PROGRESS",
}